requirements.txt
ccxt
pandas
pandas_ta
python-dotenv
numpy
requests
websocket-client
python-binance
aiohttp
